Introduce an epoch counter to ensure that CFG is explored completely
The main idea here is that the termination criterion so far was too strong. Until now, I did not process a basic block when I had seen it already from the same incoming edge.
Now, we also record the number of updates to the analysis information. So, I keep processing the basic block as long as we got information since the last time. This still guarantees termination for loops, and avoids skipping over branches we have seen previously, but not with the current set of information.

Signed-off-by: Stefan Marr <git@stefan-marr.de>